+++ src/npg_porch/endpoints/ui.py
@@ -16,12 +16,22 @@
 #
 # You should have received a copy of the GNU General Public License along with
 # this program. If not, see <http://www.gnu.org/licenses/>.
-import logging
-
+from datetime import datetime
+from enum import Enum
 from fastapi import APIRouter, Depends, Request
 from starlette import status
 
 from npg_porch.db.connection import get_DbAccessor
+from npg_porch.models import TaskStateEnum
+
+
+class UiStateEnum(str, Enum):
+    def __str__(self):
+        return self.value
+
+    ALL = "All"
+    NOT_DONE = "NOT DONE"
+
 
 router = APIRouter(
     prefix="/ui",
@@ -33,25 +43,44 @@ router = APIRouter(
 
 
 @router.get(
-    "/tasks",
+    "/tasks/{pipeline_name}/{state}/{since}",
     response_model=dict,
-    summary="Returns all expanded tasks in a displayable format for the ui",
+    summary="Returns all expanded tasks for the specified pipeline and status "
+    "in a displayable format for the ui",
 )
-async def get_ui_tasks(request: Request, db_accessor=Depends(get_DbAccessor)) -> dict:
+async def get_ui_tasks(
+    request: Request,
+    pipeline_name: str,
+    state: TaskStateEnum | UiStateEnum,
+    since: datetime,
+    db_accessor=Depends(get_DbAccessor),
+) -> dict:
+    pipeline_name = None if pipeline_name == "All" else pipeline_name
     params = request.query_params.get
-    task_list = await db_accessor.get_expanded_tasks()
+    since = None if since == datetime.min else since  # This is not perfect, but
+    state = (
+        [
+            taskstate
+            for taskstate in TaskStateEnum
+            if taskstate not in [TaskStateEnum.DONE, TaskStateEnum.CANCELLED]
+        ]
+        if state == UiStateEnum.NOT_DONE
+        else None
+        if state == UiStateEnum.ALL
+        else [state]
+    )
+    task_list = await db_accessor.get_expanded_tasks(pipeline_name, state, since)
     return {"draw": params("draw"), "recordsTotal": len(task_list), "data": task_list}
 
 
 @router.get(
-    "/tasks/{pipeline_name}",
+    "/long_running",
     response_model=dict,
-    summary="Returns all expanded tasks for the specified pipeline in a "
-    "displayable format for the ui",
+    summary="Returns all long running tasks in a displayable format for the ui",
 )
-async def get_ui_pipeline_tasks(
-    request: Request, pipeline_name: str, db_accessor=Depends(get_DbAccessor)
+async def get_long_running_ui_tasks(
+    request: Request, db_accessor=Depends(get_DbAccessor)
 ) -> dict:
     params = request.query_params.get
-    task_list = await db_accessor.get_expanded_tasks(pipeline_name)
+    task_list = await db_accessor.get_long_running_tasks()
     return {"draw": params("draw"), "recordsTotal": len(task_list), "data": task_list}